update test helpers and setup. add test for plan settings and member settings
