+++ package.json
@@ -108,7 +108,8 @@
     "typescript": "^5.3.3",
     "vite": "^5.2.6",
     "vite-tsconfig-paths": "^4.3.2",
-    "vitest": "^1.4.0"
+    "vitest": "^1.4.0",
+    "vitest-canvas-mock": "^0.3.3"
   },
   "pnpm": {
     "overrides": {
